+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+### Changed
+
+### Removed
+
+### Fixed
+
+- Binding audio elements will no longer throw an error unless calling code is
+  trying to choose an output device in a browser that does not support
+  `setSinkId`, and the demo will not log an error in these cases.
+
 ## [2.2.0] - 2020-12-04
 
 ### Added